global: addition of OAISet record adding
* Adds API method to OAISet for adding records.
